test/requirements.txt
selenium
stegano
pillow
numpy
bitarray
python-dotenv